PrjFS Mac kext: IOSharedDataQueue subclass with a fixed enqueue() implementation
A subclass of IOSharedDataQueue is added which overrides the enqueue() member function with the implementation from OS X 10.11.6, prior to the introduction of a regression which causes inconsistent memory ordering across threads.

Both user client subclasses are updated to use the new class in place of IOSharedDataQueue.

This works around the remaining occurrences of the IOSharedDataQueue stall issue which has been reported to Apple as Radar issue 43093190.